feat: add styles of garage
