+++ async-race/src/common/functions/createSvg.ts
@@ -1,7 +1,7 @@
 export function createSvg(color: string): string {
   const svg = `
 <svg version="1.0" xmlns="http://www.w3.org/2000/svg"
-width="1280.000000pt" height="867.000000pt" viewBox="0 0 1280.000000 867.000000"
+width="1280.000000pt" height="867.000000pt" viewBox="0 0 1280.000000 300.000000"
 preserveAspectRatio="xMidYMid meet">
 <metadata>
 Created by potrace 1.15, written by Peter Selinger 2001-2017
