+++ async-race/src/styles/_var.scss
@@ -1,2 +1,3 @@
 $bgcolor: black;
 $btnColor: rgb(168, 223, 168);
+$textColor: rgb(219, 219, 219);
