+++ async-race/src/styles/common.scss
@@ -11,3 +11,11 @@ body{
   background: black;
 }
 
+ul, li{
+  list-style-type: none;
+}
+
+input{
+  margin-right: 0;
+}
+
