+++ async-race/src/styles/header.scss
@@ -27,7 +27,7 @@
     }
 
     &:first-child {
-        margin-right: 1.5rem;
+        margin-right: 1rem;
     }
 }
