+++ async-race/src/styles/style.scss
@@ -1,3 +1,4 @@
 @import url('https://fonts.googleapis.com/css2?family=Maven+Pro&display=swap');
 @import './common.scss';
-@import './header.scss'
+@import './header.scss';
+@import './garage.scss';
